add Advapi32UtilFFM, Kernel32UtilFFM
