+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ##### New Features
 * [#2925](https://github.com/oshi/oshi/pull/2925): Introduced oshi-core-java25 module that intends to provide FFM support - [@rohan-coder02](https://github.com/rohan-coder02).
 * [#2949](https://github.com/oshi/oshi/pull/2949): Implement MacOperatingSystem using FFM - [@dbwiddis](https://github.com/dbwiddis).
-* [#2959](https://github.com/oshi/oshi/pull/2925): Implement WindowsOperatingSystem using FFM - [@rohan-coder02](https://github.com/rohan-coder02).
+* [#2959](https://github.com/oshi/oshi/pull/2959): Implement WindowsOperatingSystem using FFM - [@rohan-coder02](https://github.com/rohan-coder02).
+* [#2968](https://github.com/oshi/oshi/pull/2968): Add isElevated, getSystemBootTime, getThreadCount, getThreadId to WindowsOperatingSystemFFM - [@rohan-coder02](https://github.com/rohan-coder02).
 
 ##### Bug fixes / Improvements
 * [#2946](https://github.com/oshi/oshi/pull/2946): Add Hyper-V VM mac address - [@chunzhennn](https://github.com/chunzhennn).
